Made changes to accommodate for the latest version of Xcode.
Class Weather needed a different way to bridge from NSNumber to float.

The City class now has hidden parameters, to shorten the length of the code.

The colors of the text in the CompareWeather view were changed to red and blue, to visualize
which staples belong to which city.

Also wrote documentation for method viewDidLoad in class ViewController.